% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anacuali.R
\name{anacuali}
\alias{anacuali}
\title{Analisis qualitativa do IMC corporal}
\usage{
anacuali(x)
}
\arguments{
\item{x}{Valor do IMC corporal}
}
\value{
Analisis qualitativa do IMC em funcao do limite = 22
}
\description{
Analisis qualitativa do IMC corporal
}
\examples{

medidas$index <- anacuali(medidas$imc)

}
