% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{bifurcation_ts}
\alias{bifurcation_ts}
\title{Title}
\usage{
bifurcation_ts(
  model,
  model_pars,
  bifpar_list,
  X0 = c(),
  X_names = names(X0),
  seed_nr = 123,
  times = seq(0, 100, by = 0.01),
  method = c("lsoda", "euler", "rk4")[1],
  stopifregime = function(out) {
     FALSE
 },
  do_downsample = TRUE,
  downsample_pars = list(type = c("average", "one_sample")[1], win_size = 10, which_X =
    c(10, "first", "middle", "last", "random")[1], seed = 123),
  silent = FALSE
)
}
\arguments{
\item{stopifregime}{}
}
\description{
Title
}
