% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutescatterplot.R
\name{cute_scatterplot}
\alias{cute_scatterplot}
\title{super cute scatterplot}
\usage{
cute_scatterplot(data, var12, var2, lmline = FALSE)
}
\arguments{
\item{lmline}{TRUE is you want to add linear line but default is FALSE}
}
\description{
super cute scatterplot
}
\examples{
cute_scatterplot(data=iris, var1=Sepal.Width, var2=Sepal.Length, lmline=FALSE)
}
