% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downsample}
\alias{downsample}
\title{Title}
\usage{
downsample(
  df,
  X_names,
  type = c("average", "one_sample")[1],
  win_size = 10,
  which_X = c(10, "first", "middle", "last", "random")[1],
  seed_nr = 123
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
