% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{find_WCSS_per_k}
\alias{find_WCSS_per_k}
\title{Find best matching period by partitioning the timeseries into temporal clusters and minimizing within-cluster sum of squares for peak coordinates and peak indices}
\usage{
find_WCSS_per_k(ks, coord, peak_idx)
}
\arguments{
\item{peak_idx}{}
}
\description{
Find best matching period by partitioning the timeseries into temporal clusters and minimizing within-cluster sum of squares for peak coordinates and peak indices
}
