% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupByDate.R
\name{groupByDate}
\alias{groupByDate}
\title{Group Dataframe By Date}
\usage{
groupByDate(data_frame)
}
\arguments{
\item{data_frame}{A data.frame object with two columns: date and value.}
}
\value{
Return a tibble data.frame.
}
\description{
Return the mean value grouped by date.
}
\seealso{
\code{\link{group_by}}, \code{\link{summarise}}
}
