% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R
\name{plot.mvnpdf}
\alias{plot.mvnpdf}
\title{Plot of the mvnpdf function}
\usage{
\method{plot}{mvnpdf}(x, ...)
}
\arguments{
\item{x}{an object of class \code{mvnpdf} resulting from a call of
\code{mnvpdf()} function.}

\item{...}{graphical parameters passed to \code{plot()} function.}
}
\value{
Nothing is returned, only a plot is given.
}
\description{
Plot of the mvnpdf function
}
\examples{
pdfvalues <- mvnpdf(x=matrix(seq(-3, 3, by = 0.1), nrow = 1), Log=FALSE)
plot(pdfvalues)
}
