% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdf.R, R/mvnpdfoptim.R, R/mvnpdfsmart.R
\name{mvnpdf}
\alias{mvnpdf}
\alias{mvnpdfoptim}
\alias{mvnpdfsmart}
\title{Title: Densite d'une loi normale multivariees
Description: Cette fonction permet de calculer la densite d'une loi normale multivariee
Details: ceci precise plus de details sur la fonction}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)),
  Log = TRUE)

mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)),
  Log = TRUE)
}
\arguments{
\item{x}{matrice a n colonnes (observations) et p lignes}

\item{mean}{vecteur des moyennes}

\item{varcovM}{matrice des variances covariances}

\item{Log}{parametre logique parend par defaut la valeur NULL}
}
\value{
Cette fonction retourne en sortie la matrice x et la matrice y des images de x par la fonction
}
\description{
Title: Densite d'une loi normale multivariees
Description: Cette fonction permet de calculer la densite d'une loi normale multivariee
Details: ceci precise plus de details sur la fonction
}
\examples{
mvnpdf(x=matrix(1.96),Log=FALSE)
dnorm(1.96)
}
